Data.Geospatial.Internal.CRS

Description

See Section 3 Coordinate Reference System Objects in the GeoJSON Spec

Types

data CRSObject

See Section 3 Coordinate Reference System Objects in the GeoJSON Spec NoCRS is required because no crs attribute in a GeoJSON feature is NOT the same thing as a null crs attribute. no crs value implies the default CRS, while a null CRS means you cannot assume a CRS, null will mapped to NoCRS while a non-existent attribute will be mapped to a Nothing Maybe value

Constructors

NoCRS

NamedCRS !Name

EPSG Code

LinkedCRS !Href !FormatString

Functions

defaultCRS :: CRSObject

The default CRS according to Section 3 Coordinate Reference System Objects is WGS84 which I believe, from http://spatialreference.org/ref/epsg/4326/ which translates to this in JSON: http://spatialreference.org/ref/epsg/4326/json/) is represented thus:

Prisms

_NoCRS :: Prism' CRSObject ()

_NamedCRS :: Prism' CRSObject Name

_EPSG :: Prism' CRSObject Code

_LinkedCRS :: Prism' CRSObject (Href, FormatString)