Data.Geospatial.Internal.GeoPosition

Description

see Section 2.1.1 Position in the GeoJSON Spec

Type

data GeoPosition

see Section 2.1.1 Position in the GeoJSON Spec, I make the assumption here that the only position types we will use will involve easting or northing (+ve or -ve Altitude) or lon or lat (+ve or -ve Altitude)

Constructors

LonLat Longitude Latitude

LonLatAlt Longitude Latitude Altitude

EastingNorthing Easting Northing

EastingNorthingAlt Easting Northing Altitude

Functions

stripCRSFromPosition :: GeoPosition -> GeoPositionWithoutCRS

the GeoPosition is a bit special in that when you convert it to GeoJSON, it will lose the CRS info attached to it and cannot be read back in from the GeoJSON. Hence it is ineligible for the FromJSON type class, so this function will strip it down to a GeoPositionWithoutCRS, which is eligible