Data.LineString

Description

Refer to the GeoJSON Spec http://geojson.org/geojson-spec.html#linestring

A LinearString is a List with at least 2 elements

Type

data LineString a

a LineString has at least 2 elements

data ListToLineStringError

When converting a List to a LineString, here is a list of things that can go wrong:

The list was empty

The list only had one element

Constructors

ListEmpty

SingletonList

data SequenceToLineStringError

When converting a Sequence to a LineString, here is a list of things that can go wrong:

The sequence was empty

The sequence only had one element

Constructors

SequenceEmpty

SingletonSequence

Functions

toSeq :: LineString a -> Seq a

create a sequence from a LineString. LineString 1 2 [3,4] --> Sequence [1,2,3,4]

combineToSeq :: (a -> a -> b) -> LineString a -> Seq b

create a sequence from a LineString by combining values. LineString 1 2 3,4 --> Sequence [(1,2),(2,3),(3,4)]

fromSeq :: Validate v => Seq a -> v SequenceToLineStringError (LineString a)

creates a LineString out of a sequence of elements, if there are enough elements (needs at least 2) elements

fromLineString :: LineString a -> [a]

This function converts it into a list and appends the given element to the end.

fromList :: Validate v => [a] -> v ListToLineStringError (LineString a)

creates a LineString out of a list of elements, if there are enough elements (needs at least 2) elements

makeLineString

Arguments

:: a
The first element

-> a
The second element

-> Seq a
The rest of the optional elements

-> LineString a

Creates a LineString makeLineString x y zs creates a LineString homomorphic to the list [x, y] ++ zs

lineStringHead :: LineString a -> a

returns the element at the head of the string

lineStringLast :: LineString a -> a

returns the last element in the string

lineStringLength :: LineString a -> Int

returns the number of elements in the list, including the replicated element at the end of the list.