Data.Geospatial

Description

Refer to the GeoJSON Spec http://www.geojson.org/geojson-spec.html

Types

type Latitude = Double

type Longitude = Double

type Easting = Double

type Northing = Double

type Altitude = Double

data FeatureID

Constructors

FeatureIDText Text

FeatureIDNumber Int

data GeoPositionWithoutCRS

Constructors

GeoEmpty

GeoPointXY PointXY

GeoPointXYZ PointXYZ

GeoPointXYZM PointXYZM

data GeoPosition

see Section 2.1.1 Position in the GeoJSON Spec, I make the assumption here that the only position types we will use will involve easting or northing (+ve or -ve Altitude) or lon or lat (+ve or -ve Altitude)

Constructors

LonLat Longitude Latitude

LonLatAlt Longitude Latitude Altitude

EastingNorthing Easting Northing

EastingNorthingAlt Easting Northing Altitude

newtype GeoPoint

Constructors

GeoPoint

Fields

_unGeoPoint :: GeoPositionWithoutCRS

retrieveXY :: GeoPositionWithoutCRS -> PointXY

data PointXY

(GeoPositionWithoutCRS is a catch all for indeterminate CRSs and for expression of positions before a CRS has been determined

Constructors

PointXY

Fields

_xyX :: !Double

_xyY :: !Double

data PointXYZ

Constructors

PointXYZ

Fields

_xyzX :: !Double

_xyzY :: !Double

_xyzZ :: !Double

data PointXYZM

Constructors

PointXYZM

Fields

_xyzmX :: !Double

_xyzmY :: !Double

_xyzmZ :: !Double

_xyzmM :: !Double

newtype GeoMultiPoint

Constructors

GeoMultiPoint

Fields

_unGeoMultiPoint :: Seq GeoPositionWithoutCRS

splitGeoMultiPoint :: GeoMultiPoint -> Seq GeoPoint

Split GeoMultiPoint coordinates into multiple GeoPoints

mergeGeoPoints :: Seq GeoPoint -> GeoMultiPoint

Merge multiple GeoPoints into one GeoMultiPoint

newtype GeoPolygon

Constructors

GeoPolygon

Fields

_unGeoPolygon :: Seq (LinearRing GeoPositionWithoutCRS)

newtype GeoMultiPolygon

Constructors

GeoMultiPolygon

Fields

_unGeoMultiPolygon :: Seq (Seq (LinearRing GeoPositionWithoutCRS))

splitGeoMultiPolygon :: GeoMultiPolygon -> Seq GeoPolygon

Split GeoMultiPolygon coordinates into multiple GeoPolygons

mergeGeoPolygons :: Seq GeoPolygon -> GeoMultiPolygon

Merge multiple GeoPolygons into one GeoMultiPolygon

newtype GeoLine

Constructors

GeoLine

Fields

_unGeoLine :: LineString GeoPositionWithoutCRS

newtype GeoMultiLine

Constructors

GeoMultiLine

Fields

_unGeoMultiLine :: Seq (LineString GeoPositionWithoutCRS)

splitGeoMultiLine :: GeoMultiLine -> Seq GeoLine

Split GeoMultiLine coordinates into multiple GeoLines

mergeGeoLines :: Seq GeoLine -> GeoMultiLine

Merge multiple GeoLines into one GeoMultiLine

data GeospatialGeometry

See section 2.1 Geometry Objects in the GeoJSON Spec.

Constructors

NoGeometry

Point GeoPoint

MultiPoint GeoMultiPoint

Polygon GeoPolygon

MultiPolygon GeoMultiPolygon

Line GeoLine

MultiLine GeoMultiLine

Collection (Seq GeospatialGeometry)

type Name = Text

type Code = Int

type Href = Text

type FormatString = Text

type ProjectionType = Text

data CRSObject

See Section 3 Coordinate Reference System Objects in the GeoJSON Spec NoCRS is required because no crs attribute in a GeoJSON feature is NOT the same thing as a null crs attribute. no crs value implies the default CRS, while a null CRS means you cannot assume a CRS, null will mapped to NoCRS while a non-existent attribute will be mapped to a Nothing Maybe value

Constructors

NoCRS

NamedCRS !Name

EPSG Code

LinkedCRS !Href !FormatString

data BoundingBoxWithoutCRS

See Section 4 Bounding Boxes of the GeoJSON spec, The length of the list/array must be 2*n where n is the dimensionality of the position type for the CRS with min values first followed by the max values, wich both the min/max sets following the same axis order as the CRS, e.g for WGS84: minLongitude, minLatitude, maxLongitude, maxLatitude The spec mentions that it can be part of a geometry object too but doesnt give an example, This implementation will ignore bboxes on Geometry objects, they can be added if required.

Constructors

BoundingBoxWithoutCRSXY PointXY PointXY

BoundingBoxWithoutCRSXYZ PointXYZ PointXYZ

BoundingBoxWithoutCRSXYZM PointXYZM PointXYZM

data GeoFeature a

See Section 2.2 Feature Objects of the GeoJSON spec. Parameterised on the property type

Constructors

GeoFeature

Fields

_bbox :: Maybe BoundingBoxWithoutCRS

_geometry :: GeospatialGeometry

_properties :: a

_featureId :: Maybe FeatureID

reWrapGeometry :: GeoFeature a -> GeospatialGeometry -> GeoFeature a

data GeoFeatureCollection a

See Section 2.3 Feature Collection Objects of the GeoJSON spec

Constructors

GeoFeatureCollection

Fields

_boundingbox :: Maybe BoundingBoxWithoutCRS

_geofeatures :: Seq (GeoFeature a)

Functions

stripCRSFromPosition :: GeoPosition -> GeoPositionWithoutCRS

the GeoPosition is a bit special in that when you convert it to GeoJSON, it will lose the CRS info attached to it and cannot be read back in from the GeoJSON. Hence it is ineligible for the FromJSON type class, so this function will strip it down to a GeoPositionWithoutCRS, which is eligible

defaultCRS :: CRSObject

The default CRS according to Section 3 Coordinate Reference System Objects is WGS84 which I believe, from http://spatialreference.org/ref/epsg/4326/ which translates to this in JSON: http://spatialreference.org/ref/epsg/4326/json/) is represented thus:

Lenses

Geometry Lenses

unGeoPoint :: Iso' GeoPoint GeoPositionWithoutCRS

unGeoMultiPoint :: Iso' GeoMultiPoint (Seq GeoPositionWithoutCRS)

unGeoPolygon :: Iso' GeoPolygon (Seq (LinearRing GeoPositionWithoutCRS))

unGeoLine :: Iso' GeoLine (LineString GeoPositionWithoutCRS)

unGeoMultiLine :: Iso' GeoMultiLine (Seq (LineString GeoPositionWithoutCRS))

unGeoMultiPolygon :: Iso' GeoMultiPolygon (Seq (Seq (LinearRing GeoPositionWithoutCRS)))

Feature Lenses

bbox :: forall a. Lens' (GeoFeature a) (Maybe BoundingBoxWithoutCRS)

geometry :: forall a. Lens' (GeoFeature a) GeospatialGeometry

properties :: forall a a. Lens (GeoFeature a) (GeoFeature a) a a

featureId :: forall a. Lens' (GeoFeature a) (Maybe FeatureID)

boundingbox :: forall a. Lens' (GeoFeatureCollection a) (Maybe BoundingBoxWithoutCRS)

geofeatures :: forall a a. Lens (GeoFeatureCollection a) (GeoFeatureCollection a) (Seq (GeoFeature a)) (Seq (GeoFeature a))

Prisms

BasicTypes

class HasGeoPositionWithoutCRS c where

Methods

geoPositionWithoutCRS :: Lens' c GeoPositionWithoutCRS

Geometry

_NoGeometry :: Prism' GeospatialGeometry ()

_Point :: Prism' GeospatialGeometry GeoPoint

_MultiPoint :: Prism' GeospatialGeometry GeoMultiPoint

_Polygon :: Prism' GeospatialGeometry GeoPolygon

_MultiPolygon :: Prism' GeospatialGeometry GeoMultiPolygon

_Line :: Prism' GeospatialGeometry GeoLine

_MultiLine :: Prism' GeospatialGeometry GeoMultiLine

_Collection :: Prism' GeospatialGeometry (Seq GeospatialGeometry)

CRS

_NoCRS :: Prism' CRSObject ()

_NamedCRS :: Prism' CRSObject Name

_EPSG :: Prism' CRSObject Code

_LinkedCRS :: Prism' CRSObject (Href, FormatString)