Data.Geospatial.Internal.BasicTypes

Description

Basic types for GeoJSON representations.

Coordinate types

type Latitude = Double

type Longitude = Double

type Easting = Double

type Northing = Double

type Altitude = Double

data GeoPositionWithoutCRS

Constructors

GeoEmpty

GeoPointXY PointXY

GeoPointXYZ PointXYZ

GeoPointXYZM PointXYZM

retrieveXY :: GeoPositionWithoutCRS -> PointXY

data PointXY

(GeoPositionWithoutCRS is a catch all for indeterminate CRSs and for expression of positions before a CRS has been determined

Constructors

PointXY

Fields

_xyX :: !Double

_xyY :: !Double

data PointXYZ

Constructors

PointXYZ

Fields

_xyzX :: !Double

_xyzY :: !Double

_xyzZ :: !Double

data PointXYZM

Constructors

PointXYZM

Fields

_xyzmX :: !Double

_xyzmY :: !Double

_xyzmZ :: !Double

_xyzmM :: !Double

newtype DoubleArray

Constructors

DoubleArray [Double]

class HasGeoPositionWithoutCRS c where

Methods

geoPositionWithoutCRS :: Lens' c GeoPositionWithoutCRS

CRS Reference types

type Name = Text

type Code = Int

type Href = Text

type FormatString = Text

type ProjectionType = Text

Feature Types

data BoundingBoxWithoutCRS

See Section 4 Bounding Boxes of the GeoJSON spec, The length of the list/array must be 2*n where n is the dimensionality of the position type for the CRS with min values first followed by the max values, wich both the min/max sets following the same axis order as the CRS, e.g for WGS84: minLongitude, minLatitude, maxLongitude, maxLatitude The spec mentions that it can be part of a geometry object too but doesnt give an example, This implementation will ignore bboxes on Geometry objects, they can be added if required.

Constructors

BoundingBoxWithoutCRSXY PointXY PointXY

BoundingBoxWithoutCRSXYZ PointXYZ PointXYZ

BoundingBoxWithoutCRSXYZM PointXYZM PointXYZM

data FeatureID

Constructors

FeatureIDText Text

FeatureIDNumber Int